+++ src/formdefaults/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class FormdefaultsConfig(AppConfig):
+    name = "formdefaults"
+    verbose_name = "Formularze - wartości domyślne"