+++ src/formdefaults/tests/test_util.py
@@ -1,0 +1,9 @@
+from formdefaults.models import FormRepresentationManager
+from formdefaults.util import full_name
+
+
+def test_full_name():
+    assert (
+        full_name(FormRepresentationManager())
+        == "formdefaults.models.FormRepresentationManager"
+    )