+++ src/formdefaults/util.py
@@ -1,0 +1,5 @@
+def full_name(o):
+    module = o.__class__.__module__
+    if module is None or module == str.__class__.__module__:
+        return o.__class__.__name__
+    return module + "." + o.__class__.__name__